partial refactoring, fn3F0-fn3FD
