+++ ede/ede/validation_functions_facade.py
@@ -1,0 +1,37 @@
+
+#print("Y-Y"*105)
+#registro de salidas o retiros
+
+#registro de la entrega de información
+
+#registro de matrícula
+
+#integridad de datos
+from ede.ede.validation_functions.integridad_datos.fn3F0 import fn3F0
+from ede.ede.validation_functions.integridad_datos.fn3F1_ import fn3F1
+#def fn3F1(conn, return_dict):
+#    print("X-X"*105)
+#    return False
+from ede.ede.validation_functions.integridad_datos.fn3F2 import fn3F2
+from ede.ede.validation_functions.integridad_datos.fn3F3_ import fn3F3
+#def fn3F3(conn, return_dict):
+#    print("X-3"*105)
+#    return False
+from ede.ede.validation_functions.integridad_datos.fn3F4_ import fn3F4
+#def fn3F4(conn, return_dict):
+#    print("X-4"*105)
+#    return False
+from ede.ede.validation_functions.integridad_datos.fn3F5 import fn3F5
+from ede.ede.validation_functions.integridad_datos.fn3F6 import fn3F6
+from ede.ede.validation_functions.integridad_datos.fn3F7 import fn3F7
+from ede.ede.validation_functions.integridad_datos.fn3F8 import fn3F8
+from ede.ede.validation_functions.integridad_datos.fn3F9 import fn3F9
+from ede.ede.validation_functions.integridad_datos.fn3FA import fn3FA 
+from ede.ede.validation_functions.integridad_datos.fn3FB import fn3FB
+from ede.ede.validation_functions.integridad_datos.fn3FC import fn3FC
+from ede.ede.validation_functions.integridad_datos.fn3FD import fn3FD
+
+
+
+
+#libro de classes digital
